(minor) documentation changes
 upgrade_v2.sh:
- Fixed a broken anchor tag reference
- Fixed a title reference

README.md:

Replaced the "If you use the pre v2 default Archiva configuration" section with a slimmed down section titled "Enable configuration migrations" . The motivation behind this change was to keep the documentation as easy-to-follow as possible and omit details that aren't relevant to most people. If anybody wants to see how the migration works, they can always check the upgrade script, which was well commented.